# internal/app/openapi.yaml
openapi: 3.1.0
info:
  title: Kore API
  version: 1.0.0
  description: API Kore — monolithe modulaire PSA/ESN
servers:
  - url: http://localhost:8080/api/v1
paths:
...
  /auth/oidc/authorize:
    get:
      summary: URL de redirection OIDC SSO
      parameters:
        - name: tenant
          in: query
          required: true
          schema: { type: string, format: uuid }
        - name: redirect_uri
          in: query
          required: true
          schema: { type: string }
        - name: code_challenge
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: URL authorize
...
  /admin/identity-providers:
    get:
      summary: Liste IdP du tenant
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Liste
  /admin/identity-providers/{id}:
    put:
      summary: Configurer IdP OIDC
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Configuré
...
  /gantt:
    get:
      summary: Données Gantt
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: OK }
  /planning:
    get:
      summary: Planning 60j
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: OK }
  /dashboards/{code}:
    get:
      summary: Dashboard applicatif
      security: [{ bearerAuth: [] }]
      parameters: [{ name: code, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { description: OK }
  /reports/run:
    post:
      summary: Exécuter rapport
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: OK }
  /billing-stats:
    get:
      summary: Statistiques CA
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: OK }
  /missions:
    get:
      summary: Liste missions SSII
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: OK }
    post:
      summary: Créer mission
      security: [{ bearerAuth: [] }]
      responses:
        "201": { description: Créé }
  /missions/{id}:
    get:
      summary: Détail mission
      security: [{ bearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses:
        "200": { description: OK }
  /missions/{id}/stop:
    post:
      summary: Arrêter mission
      security: [{ bearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses:
        "200": { description: OK }
  /missions/{id}/end-date:
    put:
      summary: Modifier fin mission
      security: [{ bearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses:
        "200": { description: OK }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT